perf(cats/sql): Change relationship table index with rel_agent_hash to include rel_type.
Including the rel_type in the the index will improve the query where rel_agent_hash and rel_type is given. There's also a query to get all entries with a given rel_agent_hash, this index still works for that query since the rel_agent_hash is the first column.